Code review amends
